+++ src/client/actions/projectData.ts
@@ -38,3 +38,6 @@ export const updateProjectFailed = createAction("Update project failure")();
 
 export const exportCsv = createAction("Export project CSV")<IProject>();
 export const exportCsvFailure = createAction("Export project CSV failure")<string>();
+
+export const exportShp = createAction("Export project Shapefile")<IProject>();
+export const exportShpFailure = createAction("Export project Shapefile failure")<string>();